+++ hutool-core/src/main/java/cn/hutool/core/convert/impl/TimeZoneConverter.java
@@ -1,8 +1,9 @@
 package cn.hutool.core.convert.impl;
 
-import java.util.TimeZone;
-
 import cn.hutool.core.convert.AbstractConverter;
+import cn.hutool.core.date.TimeZoneUtil;
+
+import java.util.TimeZone;
 
 /**
  * TimeZone转换器
@@ -14,7 +15,7 @@ public class TimeZoneConverter extends AbstractConverter<TimeZone>{
 
 	@Override
 	protected TimeZone convertInternal(Object value) {
-		return TimeZone.getTimeZone(convertToStr(value));
+		return TimeZoneUtil.getTimeZone(convertToStr(value));
 	}
 
 }